changes to Implementation

+++ src/main/java/AntiTask.java
@@ -1,16 +1,14 @@
-import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-public class AntiTask extends Task
-{
+
+public class AntiTask extends Task {
     public final String taskTypeString = "Anti";
-    public AntiTask(String name, String type, long startTime, double duration, long date)
-    {
+
+    public AntiTask(String name, String type, double startTime, double duration, long date) {
         super(name, type, startTime, duration, date);
     }
+
     @Override
-    public JSONObject toJSONObject()
-    {
+    public JSONObject toJSONObject() {
         JSONObject object = new JSONObject();
         object.put("Name", this.getName());
         object.put("Type", this.getType());
@@ -19,27 +17,27 @@ public class AntiTask extends Task
         object.put("Date", this.getStartDate());
         return object;
     }
-    public void setDate(int date)
-    {
+
+    public void setDate(int date) {
         this.setStartDate(date);
     }
-    public long getDate()
-    {
+
+    public long getDate() {
         return this.getStartDate();
     }
+
     @Override
-    public String toString()
-    {
+    public String toString() {
         return "Name: " + this.getName()
                 + ", Type: " + this.getType()
                 + ", StartTime: " + this.getStartTime()
                 + ", Duration: " + this.getDuration()
                 + ", Date " + this.getDate();
     }
+
     @Override
-    public boolean ifInThatDate(long startDay, long endDay)
-    {
-        if(this.getStartDate() >=startDay && this.getStartDate() <= endDay)
+    public boolean ifInThatDate(long startDay, long endDay) {
+        if (this.getStartDate() >= startDay && this.getStartDate() <= endDay)
             return true;
         return false;
     }
