+++ src/main/java/Task.java
@@ -1,26 +1,23 @@
-import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
 
-public class Task
-{
+public class Task {
     private String uuid;
     private String name;
     private String type;
-    private long startTime;
+    private double startTime;
     private double duration;
     private long startDate;
     public String taskTypeString = "NA";
-    public Task(String name, String type, long startTime, double duration, long startDate)
-    {
+
+    public Task(String name, String type, double startTime, double duration, long startDate) {
         this.name = name;
         this.type = type;
         this.startTime = startTime;
         this.duration = duration;
         this.startDate = startDate;
     }
-    public JSONObject toJSONObject()
-    {
+
+    public JSONObject toJSONObject() {
         return null;
     }
 
@@ -32,53 +29,54 @@ public class Task
         this.startDate = startDate;
     }
 
-    public void setName(String name)
-    {
+    public void setName(String name) {
         this.name = name;
     }
-    public void setType(String type)
-    {
+
+    public void setType(String type) {
         this.type = type;
     }
-    public void setStartTime(int time)
-    {
+
+    public void setStartTime(double time) {
         startTime = time;
     }
-    public void setDuration(int duration)
-    {
+
+    public void setDuration(int duration) {
         this.duration = duration;
     }
+
     public void setUuid(String uuid) {
         this.uuid = uuid;
     }
-    public String getName()
-    {
+
+    public String getName() {
         return name;
     }
-    public String getType()
-    {
+
+    public String getType() {
         return type;
     }
-    public long getStartTime()
-    {
+
+    public double getStartTime() {
         return startTime;
     }
-    public double getDuration()
-    {
+
+    public double getDuration() {
         return duration;
     }
+
     public String getUuid() {
         return uuid;
     }
-    public String toString()
-    {
+
+    public String toString() {
         return "Name: " + this.getName()
                 + "\nType: " + this.getType()
                 + "\nStartTime: " + this.getStartTime()
                 + "\nDuration: " + this.getDuration();
     }
-    public boolean ifInThatDate(long startDay, long endDay)
-    {
+
+    public boolean ifInThatDate(long startDay, long endDay) {
 
         return false;
     }
