+++ src/main/java/TransientTask.java
@@ -1,17 +1,14 @@
-import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
 
-public class TransientTask extends Task
-{
+public class TransientTask extends Task {
     public final String taskTypeString = "Transient";
-    public TransientTask(String name, String type, long startTime, double duration, long date)
-    {
+
+    public TransientTask(String name, String type, double startTime, double duration, long date) {
         super(name, type, startTime, duration, date);
     }
+
     @Override
-    public JSONObject toJSONObject()
-    {
+    public JSONObject toJSONObject() {
         JSONObject object = new JSONObject();
         object.put("Name", this.getName());
         object.put("Type", this.getType());
@@ -20,27 +17,27 @@ public class TransientTask extends Task
         object.put("Date", this.getStartDate());
         return object;
     }
-    public void setDate(long date)
-    {
+
+    public void setDate(long date) {
         this.setStartDate(date);
     }
-    public long getDate()
-    {
+
+    public long getDate() {
         return this.getStartDate();
     }
+
     @Override
-    public String toString()
-    {
+    public String toString() {
         return "Name: " + this.getName()
                 + ", Type: " + this.getType()
                 + ", StartTime: " + this.getStartTime()
                 + ", Duration: " + this.getDuration()
                 + ", Date " + this.getDate();
     }
+
     @Override
-    public boolean ifInThatDate(long startDay, long endDay)
-    {
-        if(this.getStartDate() >=startDay && this.getStartDate() <= endDay)
+    public boolean ifInThatDate(long startDay, long endDay) {
+        if (this.getStartDate() >= startDay && this.getStartDate() <= endDay)
             return true;
         return false;
     }
